requirements.txt
python-dotenv==1.0.0
requests==2.31.0
newsapi-python==0.2.7
transformers==4.36.2
torch==2.1.2
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0
wordcloud==1.9.3
diskcache==5.6.3
loguru==0.7.2